Added Multiple Page Support
So the commit message is kind of a lie. Instead of adding support to "scroll" through several pages of a career website, I decided to use the 'filtered" url for the websites that require scrolling through multipl pages (i.e imc). However, it didn't work right off the bat, so I had to use selenium to parse the webpage. I think this is because the url redirect is happening in the actual JS code (I could be wring tho). This made it a bit slower, but we shall see how this goes. Still can't figure out nested HTML classes tho